requirements.txt
fastapi==0.115.14
uvicorn==0.35.0
python-multipart==0.0.20
langchain==0.3.26
langchain-community==0.3.26
langchain-openai==0.3.27
langchain-chroma==0.2.4
pypdf==5.7.0
pydantic-settings==2.10.1
motor==3.7.1